--- src/test/java/tests/Calculator.java
@@ -1,7 +1,0 @@
-package tests;
-
-public class Calculator {
-    public int sum(int a, int b) {
-        return a + b;
-    }
-}
