--- src/test/java/tests/XPathSelectorTest.java
@@ -1,102 +1,0 @@
-package tests;
-
-import factory.BrowsersFactory;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.testng.Assert;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.Test;
-
-public class XPathSelectorTest {
-    private WebDriver driver;
-
-    @BeforeMethod
-    public void setUp() {
-        driver = new BrowsersFactory().getDriver();
-    }
-
-    @AfterMethod
-    public void tearDown() {
-        driver.quit();
-    }
-
-    @Test
-    public void cssXPathTest() {
-        driver.get("D:\\JavaTeachMeSkills\\TAF_AQA18_onl\\src\\test\\resources\\index.html");
-
-        //Absolut path
-        Assert.assertTrue(driver.findElement(By.xpath("/html/body/div/div/div/div[3]")).isDisplayed());
-
-        //all elements
-        //Assert.assertEquals(128, driver.findElement(By.xpath("//*")));
-
-        //searcg tagname
-        Assert.assertTrue(driver.findElement(By.xpath("//h1")).isDisplayed());
-
-        // Аналог родительского div и на один уровень ниже h1
-        Assert.assertTrue(driver.findElement(By.xpath("//div/h1")).isDisplayed());
-
-        // Аналог родительского div и на любом уровене ниже div
-        Assert.assertTrue(driver.findElement(By.xpath("//div//div")).isDisplayed());
-
-        // Поиск элемента с тэгом div у которого есть аттрибут id
-        Assert.assertTrue(driver.findElement(By.xpath("//div[@id]")).isDisplayed());
-
-        // Поиск элемента у которого есть аттрибут id cо значением top-logo
-        Assert.assertTrue(driver.findElement(By.xpath("//*[@id = 'top-logo']")).isDisplayed());
-
-        // Поиск элемента у которого есть аттрибут method cо значением и аттрибут target со значением
-        Assert.assertTrue(driver.findElement(By.xpath("//*[@method='post' and @target='_blank']")).isDisplayed());
-
-        // Поиск элемента c аттрибутом lang или class со значением
-        Assert.assertTrue(driver.findElement(By.xpath("//*[@lang or @class=\"noSel newsletter\"]")).isDisplayed());
-
-        // Поиск элемента у которого значение аттрибута начинается с
-        Assert.assertTrue(driver.findElement(By.xpath("//*[starts-with(@id, 'new')]")).isDisplayed());
-
-        // Поиск элемента у которого значение аттрибута содержит подстроку
-        Assert.assertTrue(driver.findElement(By.xpath("//div[contains(@id, 'Template')]")).isDisplayed());
-
-        // Поиск элемента у которого текстовое значение содержит подстроку
-        Assert.assertTrue(driver.findElement(By.xpath("//*[contains(text(), 'All Projects')]")).isDisplayed());
-
-        // Поиск элемента по индексу
-        Assert.assertTrue(driver.findElement(By.xpath("//div[@class = 'summary-links text-secondary']/a[2]")).isDisplayed());
-
-
-        // Поиск элемента у которого текстовое значение содержит равно
-        Assert.assertTrue(driver.findElement(By.xpath("//*[text() = 'We are all animals!']")).isDisplayed());
-        Assert.assertTrue(driver.findElement(By.xpath("//*[. = '<table>']")).isDisplayed());
-
-    }
-
-    @Test
-    public void xPathAxesTest() {
-        driver.get("D:\\JavaTeachMeSkills\\TAF_AQA18_onl\\src\\test\\resources\\index.html");
-
-        // Поиск родителя у элемента с тэгом h1
-        Assert.assertTrue(driver.findElement(By.xpath("//h1/..")).isDisplayed());
-        Assert.assertTrue(driver.findElement(By.xpath("//h1/parent::div")).isDisplayed());
-
-        // Поиск всех предков с тэгом div у элемента с тэгом h1
-        Assert.assertTrue(driver.findElement(By.xpath("//h1/ancestor::div")).isDisplayed());
-
-        // Использование child - непосредственно дочерние элементы с тэго a от div
-        Assert.assertTrue(driver.findElement(By.xpath("//div/a")).isDisplayed());
-        Assert.assertTrue(driver.findElement(By.xpath("//div/child::a")).isDisplayed());
-
-        // Использование child - все дочерние элементы с тэго a от div
-        Assert.assertTrue(driver.findElement(By.xpath("//div//a")).isDisplayed());
-        Assert.assertTrue(driver.findElement(By.xpath("//div/descendant::a")).isDisplayed());
-
-        //Использование following - Выбирает всё в документе после закрытия тэга текущего узла
-        Assert.assertTrue(driver.findElement(By.xpath("//*[@class='dialog-title']/following::form")).isDisplayed());
-
-        //Использование preceding- Выбирает все узлы, которые появляются перед текущим узлом в документе
-        Assert.assertTrue(driver.findElement(By.xpath("//*[@class=\"dialog-title\"]/preceding::form")).isDisplayed());
-
-        //Использование preceding-sibling - Выбирает все узлы одного уровня до текущего узла
-        Assert.assertTrue(driver.findElement(By.xpath("//*[@class='dialog-title']/preceding-sibling::form")).isDisplayed());
-    }
-}
