--- src/test/java/tests/date/RetryTest.java
@@ -1,22 +1,0 @@
-package tests.date;
-
-import org.testng.Assert;
-import org.testng.annotations.Test;
-import tests.BaseTest;
-import utils.Retry;
-
-public class RetryTest extends BaseTest {
-    private int attempt = 1;
-
-    @Test (retryAnalyzer = Retry.class)
-    public void flakyTest() {
-        if (attempt == 3) {
-            Assert.assertTrue(true);
-        } else {
-            attempt++;
-            System.out.println("Attempt is: " + attempt);
-            throw new NullPointerException();
-            // Assert.assertTrue(false);
-        }
-    }
-}
