--- src/test/java/tests/simpleTest1.java
@@ -1,14 +1,0 @@
-package tests;
-
-import org.testng.annotations.Test;
-
-public class simpleTest1 extends BaseTest {
-    @Test
-    public void simpleTest1(){
-        System.out.println("this is my 2_1 test!");
-    }
-    @Test
-    public void simpleTest2(){
-        System.out.println("this is my 2_2 test!");
-    }
-}
