+++ src/test/resources/config.properties
@@ -1,4 +1,4 @@
-url=https://aqa18onl01.testrail.io/
+url=https://www.saucedemo.com/
 browser=chrome
 headless=false
 username=atrostyanko@gmail.com
